requirements.txt
psutil
pylint
slackclient
python-dotenv
